pyproject.toml
[tool.poetry]
name = "data-validator"
version = "0.1.0"
description = ""
authors = ["elroy-galbraith <elroy.galbraith@aeontsolutions.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.41.1"
pandas = "^2.2.3"
google-cloud-bigquery = "^3.27.0"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
python-dotenv = "^1.0.1"
db-dtypes = "^1.3.1"
streamlit-authenticator = "^0.4.1"
authlib = "^1.4.0"
google-cloud-bigquery-storage = "^2.27.0"
pytz = "^2024.2"
folium = "^0.19.4"
streamlit-folium = "^0.24.0"
pydeck = "^0.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
